+++ src/App/components/MyPlayer/components/PlayerListItem.scss
